+++ src/components/traits/i-lock-page-scroll/i-lock-page-scroll.ts
@@ -45,7 +45,8 @@ export default abstract class iLockPageScroll {
 
 		const lockLabel = {
 			group,
-			label: $$.lock
+			label: $$.lock,
+			join: true
 		};
 
 		if (is.mobile !== false) {
@@ -122,7 +123,7 @@ export default abstract class iLockPageScroll {
 					res();
 
 				}, lockLabel);
-			}), {join: true, ...lockLabel});
+			}), lockLabel);
 
 		} else {
 			promise = $a.promise(new Promise((res) => {
@@ -141,7 +142,7 @@ export default abstract class iLockPageScroll {
 					res();
 
 				}, lockLabel);
-			}), {join: true, ...lockLabel});
+			}), lockLabel);
 		}
 
 		return promise;
@@ -213,9 +214,7 @@ export default abstract class iLockPageScroll {
 	}
 
 	/**
-	 * Locks scrolling of the document,
-	 * i.e., prevents any scrolling of the document except within the specified node.
-	 *
+	 * Locks scrolling of the document, preventing any scrolling of the document except within that node
 	 * @param [_scrollableNode] - the node within which scrolling is allowed
 	 */
 	lockPageScroll(_scrollableNode?: Element): Promise<void> {