+++ src/core/component/state/CHANGELOG.md
@@ -9,6 +9,16 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.??.?? (2022-??-??)
+
+#### :rocket: New Feature
+
+* Exposes methods to `set`/`unset` new set properties and `watch` for them changes
+
+#### :memo: Documentation
+
+* Added the normal documentation
+
 ## v3.0.0-rc.109 (2020-12-15)
 
 #### :bug: Bug Fix