+++ src/core/component/state/CHANGELOG.md
@@ -9,7 +9,7 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
-## v3.??.?? (2022-??-??)
+## v4.0.0-alpha.1 (2022-12-14)
 
 #### :rocket: New Feature